gst-omx: Re-enable output port after transition to idle state
In gst-omx, when new format is detected the first time, output port
is disabled. Then the component state is set to Idle and after that
to executing. The output port is not enabled at a later point by
gst-omx, so the Intel OMX component cannot write decoded frames into
the output port. This patch re-enables the output port.

Signed-off-by skibey <sameer.kibey@intel.com>

Change-Id: I0f75e5915c54a33f64c2b00628d0eca377f06339